Branched rebased to fetch Travis-CI from upstream
