+++ src/main/java/com/chaintope/openassetsj/utils/Utils.java
@@ -24,171 +24,12 @@ import com.google.common.primitives.Bytes;
  */
 public class Utils {
 
-	private static final String network = "testnet";
-    private static NetworkParameters params = TestNet3Params.get();
-//	private static final String network = "mainnet";
-//    private static NetworkParameters params = MainNetParams.get();
-
-    private static final int OA_NAMESPACE = 19;
-
-    private static final int OA_VERSION_BYTE = 23;
-    private static final int OA_VERSION_BYTE_TESTNET = 115;
-
-    /**
-     * Generates OpenAsset address using EC key
-     * @param ecKey ECKey from which you have to create OpenAsset Address
-     * @return OpenAsset address
-     */
-    public static String generateOaAddress(ECKey ecKey) {
-
-        Address address = ecKey.decompress().toAddress(params);
-        String oaAddress = addressToOaAddress(address);
-        
-        return oaAddress;
-    }
-
-    /**
-     * Generates OpenAsset address using uncompressed address
-     * @param address Address from which OpenAsset Address should be generated
-     * @return OpenAsset address
-     */
-    public static String addressToOaAddress(Address address) {
-
-        String oaAddress = "";
-        
-        try {
-            byte[] pubkeyHash = address.getHash160();
-            int addrLen = 1 + 20;
-            byte[] data = new byte[addrLen];
-            System.arraycopy(pubkeyHash, 0, data, 1, 20);
-            data[0] = (byte) address.getVersion();
-            oaAddress = new VersionedChecksummedBytes(OA_NAMESPACE, data) { }.toString();
-        }
-        catch (Exception e) {
-
-            e.printStackTrace();
-        }
-        return oaAddress;
-    }
-    
-    /**
-     * Converts OpenAsset address to normal bitcoin address
-     * @param oaAddress openassets address
-     * @return bitcoin address
-     */
-    public static Address oaAddressToAddress(String oaAddress) {
-
-        byte[] decodedAddress = Base58.decodeChecked(oaAddress);
-        byte[] normalAddress = new byte[20];
-        /* First 2 bytes - OAVersion byte and address version byte - skipped */
-        System.arraycopy(decodedAddress, 2, normalAddress , 0, 20);
-        Address newAddress = new Address(params, normalAddress);
-        return newAddress;
-    }
-
-    /**
-     * Generates OpenAsset ID from ECKey
-     * @param ecKey Key used to derive asset id
-     * @return Asset ID
-     */
-    public static String generateAssetID(ECKey ecKey) {
-
-        Address address = ecKey.decompress().toAddress(params);
-        byte[] pubKeyHash = address.getHash160();
-        String assetId = pubKeyHashToAssetID(pubKeyHash);
-        return assetId;
-    }
-
-    /**
-     * Generates OpenAsset ID from Public key hash
-     * @param pubKeyHashStr Public key hash used to derive asset id
-     * @return Asset ID
-     */
-    public static String pubKeyHashToAssetID(String pubKeyHashStr) {
-
-        byte[] pubKeyHash = packHexStringToBytes (pubKeyHashStr);
-        String assetId = pubKeyHashToAssetID(pubKeyHash);
-        return assetId;
-    }
-
-    /**
-     * Generates OpenAsset ID from Public key hash bytes
-     * @param pubKeyHash Public key hash used to derive asset id
-     * @return Asset ID
-     */
-    public static String pubKeyHashToAssetID(byte[] pubKeyHash) {
-
-        byte[] script = new byte[5 + pubKeyHash.length];
-        script[0] = (byte) ScriptOpCodes.OP_DUP;
-        script[1] = (byte) ScriptOpCodes.OP_HASH160;
-        script[2] = (byte) 0x14; // 20
-        System.arraycopy(pubKeyHash, 0, script, 3, pubKeyHash.length);
-        script[script.length - 2] = (byte) ScriptOpCodes.OP_EQUALVERIFY;
-        script[script.length - 1] = (byte) ScriptOpCodes.OP_CHECKSIG;
-
-        String assetId = scriptToAssetID(script);
-        return assetId;
-    }
-
-    /**
-     * Generates OpenAsset ID from script
-     * @param script Script bytes used to derive asset id
-     * @return Asset ID
-     */
-    public static String scriptToAssetID(String scriptStr) {
-    	
-        byte[] script = packHexStringToBytes (scriptStr);
-        String assetId = scriptToAssetID(script);
-        return assetId;
-    }
-    
-    /**
-     * Generates OpenAsset ID from script bytes
-     * @param script Script bytes used to derive asset id
-     * @return Asset ID
-     */
-    public static String scriptToAssetID(byte[] script) {
-
-        byte[] scriptHash160 = hash160(script);
-
-        byte[] scriptWithVersionByte = new byte[scriptHash160.length + 1];
-        scriptWithVersionByte[0] = (byte) getVersionByte();
-        System.arraycopy(scriptHash160, 0, scriptWithVersionByte, 1, scriptHash160.length);
-
-        byte[] checksum = checksum(scriptWithVersionByte);
-
-        byte[] assetIdBytes = new byte[scriptWithVersionByte.length + 4]; // 4 bytes for checksum
-        System.arraycopy(scriptWithVersionByte, 0, assetIdBytes, 0, scriptWithVersionByte.length);
-        System.arraycopy(checksum, 0, assetIdBytes, scriptWithVersionByte.length, 4);
-
-        return Base58.encode(assetIdBytes);
-    }
-    
-    /**
-     * Gets the version byte according to the current network parameters
-     */
-    private static int getVersionByte() {
-
-    	int version;
-    	switch(network) {
-
-	    	case "testnet":
-	    		version = OA_VERSION_BYTE_TESTNET;
-	    		break;
-	    	case "mainnet":
-	    	default:
-	    		version = OA_VERSION_BYTE;
-	    		break;
-		}    	
-    	return version;
-    }
-
     /**
      * Hash160 of the given input byte array
      * @param bytes Bytes to be hashed
      * @return Hashed bytes
      */
-    private static byte[] hash160(byte[] bytes) {
+    public static byte[] hash160(byte[] bytes) {
 
         SHA256Digest firstDigest = new SHA256Digest();
         firstDigest.update(bytes, 0, bytes.length);
@@ -206,7 +47,7 @@ public class Utils {
      * @param arr Byte array to calculate checksum
      * @return Checksum bytes
      */
-    private static byte[] checksum(byte[] arr) {
+    public static byte[] checksum(byte[] arr) {
 
         byte[] checksum = null;
         try {
@@ -290,106 +131,6 @@ public class Utils {
             unpackedData += Integer.toHexString((int) packedStr.charAt(i));
         }
         return unpackedData;
-    }
-
-    /**
-     * Converts input value to unsigned leb128 equivalent
-     * @param value integer value to convert into unsigned leb128
-     * @return unsigned leb128
-     */
-    public static String encodeLeb128(int value) {
-
-        String binaryStr = Integer.toString(value, 2);
-        int strLength = (((int) Math.floor(binaryStr.length() / 7)) * 7);
-        strLength += (strLength == binaryStr.length() ? 0 : 7);
-        String paddedBinaryString = String.format("%" + strLength + "s", binaryStr).replace(' ', '0');
-        int loopCount = paddedBinaryString.length() / 7;
-        List<String> lebBytes = new ArrayList<String>();
-        for (int i = 0; i < loopCount; i++) {
-
-            String paddedStr = ( i == 0 ? "0" : "1") + paddedBinaryString.substring(i*7, (i+1)*7);
-            int decimal = Integer.parseInt(paddedStr, 2);
-            String hexStr = getHexString(decimal);
-            lebBytes.add(hexStr);
-        }
-        Collections.reverse(lebBytes);
-        String leb128Str = Joiner.on("").join(lebBytes.toArray());
-        return leb128Str;
-    }
-
-    /**
-     * Reads data from LEB128 encoded string
-     * @param payload
-     * @param offset
-     * @return List containing quantity and length
-     */
-    public static List<Object> readLeb128(String payload, int offset) {
-
-            List<Object> objList = new ArrayList<>();
-        Object quantityObj = null;
-        Object lengthObj = 0;
-
-        byte[] data = Utils.packHexStringToBytes(payload);
-
-        int result = 0;
-        int shift = 0;
-
-        while(true) {
-
-            if (data.length < (1 + offset)) {
-
-                quantityObj = null;
-                lengthObj = offset;
-            }
-
-            byte byteToCheck = data[offset];
-            result = result | ((byteToCheck & 0x7F) << shift);
-
-            if (0 == (int)(byteToCheck & 0x80)) {
-
-                break;
-            }
-            shift += 7;
-            offset += 1;
-        }
-
-        quantityObj = result;
-        lengthObj = offset + 1;
-
-        objList.add(quantityObj);
-        objList.add(lengthObj);
-
-        return objList;
-    }
-
-    /**
-     * Decodes the LEB128 encoded string
-     * @param encodedString
-     * @return LEB128 decoded number
-     */
-    public static long decodeLeb128(String encodedString) {
-
-        byte[] data = Utils.packHexStringToBytes(encodedString);
-
-        long result = 0;
-        int shift = 0;
-        int offset = 0;
-
-        while (true) {
-
-            byte byteToCheck = data[offset];
-
-            result = result | ((byteToCheck & 0x7F) << shift);
-
-            if (0 == (int) (byteToCheck & 0x80)) {
-
-                break;
-            }
-            shift += 7;
-            offset += 1;
-        }
-
-        return result;
     }
 
     /**
